refactor: move `MethodCall` utility to separate module
