+++ bevy_lint/src/lints/restriction/panicking_methods.rs
@@ -59,7 +59,10 @@ use rustc_span::Symbol;
 
 use crate::{
     declare_bevy_lint, declare_bevy_lint_pass,
-    utils::hir_parse::{MethodCall, generic_args_snippet, span_args},
+    utils::{
+        hir_parse::{generic_args_snippet, span_args},
+        method_call::MethodCall,
+    },
 };
 
 declare_bevy_lint! {
