+++ bevy_lint/src/lints/suspicious/unit_in_bundle.rs
@@ -63,7 +63,9 @@ use rustc_span::Symbol;
 )]
 use rustc_type_ir::PredicatePolarity;
 
-use crate::{declare_bevy_lint, declare_bevy_lint_pass, paths, sym, utils::hir_parse::MethodCall};
+use crate::{
+    declare_bevy_lint, declare_bevy_lint_pass, paths, sym, utils::method_call::MethodCall,
+};
 
 declare_bevy_lint! {
     pub(crate) UNIT_IN_BUNDLE,
