+++ bevy_lint/src/utils/mod.rs
@@ -2,3 +2,4 @@
 
 pub mod cargo;
 pub mod hir_parse;
+pub mod method_call;
